[26.2] Fix config loading on game-dir-less environments; relocate junit tests
Real-world verification (full Gradle build, real dedicated-server boots and
the JUnit FML harness) surfaced three issues, now fixed:

- LoadingConfig was never read from disk: the static instance was
  pre-initialized to defaults, so load() (which reads/writes
  neoforge-accel.json and honors --perf) was never invoked. The instance now
  starts null and is initialized once by CommonModLoader.
- The JUnit harness boots FML without a usable game directory (null/empty
  path). LoadingConfig, ModIndexCache, CompatPrecheck, LoadingPerf and
  DiagnosticsCollector now fall back to defaults and skip file I/O instead of
  NPEing on null parents.
- Junit tests lived in packages that exist in the neoforge module
  (net.neoforged.neoforge.loading.*), causing a split-package error in the
  JPMS test modules. They were moved to net.neoforged.neoforge.unittest and
  the needed internal members (CompatPrecheck.run, ModIndexCache.load/entries)
  were made public.

Verified with the real toolchain: ./gradlew :neoforge:setup, checkFormatting
(immaculate + license) and :tests:junitTest all pass, and :tests:runServer
boots a dedicated server where the precheck analyzes 108 mods, writes
neoforge-compat-report.json + neoforge-mod-index.json + neoforge-accel.json,
and a repeat boot reuses the persisted analysis (cache: 3 hit(s)/1 miss(es)).
