--- tests/src/junit/java/net/neoforged/neoforge/loading/cache/package-info.java
@@ -1,9 +1,0 @@
-/*
- * Copyright (c) NeoForged and contributors
- * SPDX-License-Identifier: LGPL-2.1-only
- */
-
-@NullMarked
-package net.neoforged.neoforge.loading.cache;
-
-import org.jspecify.annotations.NullMarked;
