+++ tests/src/junit/java/net/neoforged/neoforge/unittest/BreakingChangesDatabaseTest.java
@@ -3,11 +3,12 @@
  * SPDX-License-Identifier: LGPL-2.1-only
  */
 
-package net.neoforged.neoforge.loading.adapt;
+package net.neoforged.neoforge.unittest;
 
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import net.neoforged.neoforge.loading.adapt.BreakingChangesDatabase;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
