+++ tests/src/junit/java/net/neoforged/neoforge/unittest/ModIndexCacheTest.java
@@ -3,12 +3,13 @@
  * SPDX-License-Identifier: LGPL-2.1-only
  */
 
-package net.neoforged.neoforge.loading.cache;
+package net.neoforged.neoforge.unittest;
 
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.Map;
+import net.neoforged.neoforge.loading.cache.ModIndexCache;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
